Create Resume_Sorting
This project is helpful for the HR department in IT as they have received a huge amount of Resumes on the daily basis. This python code will sort the resumes for the further interview process if it contains the required skill for the open positions. It will save time for HR by doing this manually as this is an automatic process and time-saving as well.